requirements.txt
asgiref==3.5.2
Django==4.1.2
django-calculation==1.0.0
django-crispy-forms==1.14.0
django-otp==1.1.3
django-widget-tweaks==1.4.12
psycopg2-binary==2.9.5
sqlparse==0.4.3
